backend/requirements.txt
fastapi
uvicorn
pandas
scipy
statsmodels
supabase
python-dotenv
python-multipart
pg8000
